Cleared unnecessary files
