+++ app/models/drive.rb
@@ -8,4 +8,16 @@ class Drive < ApplicationRecord
   has_many :candidates, through: :drives_candidates
   has_and_belongs_to_many :problems
   has_one :rule
+
+  def yet_to_start?
+    return true if start_date > Time.current
+  end
+
+  def ended?
+    return true if end_date < Time.current
+  end
+
+  def ongoing?
+    return true if start_date < Time.current && Time.current < end_date
+  end
 end
