--- spec/factories/organizations.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-FactoryBot.define do
-  factory :organization, class: Organization do
-  end
-end
